pyproject.toml
[tool.poetry]
name = "metac-bot-template"
version = "0.1.0"
description = ""
authors = ["Vasile Popescu <elisescu@elisescu.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
requests = "^2.32.3"
numpy = "^2.3.0"
openai = "^2.0.0"
python-dotenv = "^1.0.1"
pydantic = "^2.0.0"
exa-py = "^1.0.0"
firecrawl-py = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

